# Conflicts:
#	app/src/main/java/team/y2k2/globa/api/ApiService.java
#	app/src/main/java/team/y2k2/globa/api/model/entity/Keyword.java
#	app/src/main/java/team/y2k2/globa/api/model/request/WithdrawRequest.java
#	app/src/main/java/team/y2k2/globa/docs/more/DocsMoreActivity.java
#	app/src/main/java/team/y2k2/globa/docs/quiz/quizconduct/QuizActivity.java
#	app/src/main/java/team/y2k2/globa/main/folder/edit/FolderNameEditActivity.java
#	app/src/main/java/team/y2k2/globa/main/folder/inside/FolderInsideFragment.java
#	app/src/main/java/team/y2k2/globa/main/folder/share/FolderShareActivity.java
#	app/src/main/java/team/y2k2/globa/main/profile/ProfileFragment.java

+++ app/src/main/java/team/y2k2/globa/api/model/entity/Keyword.java
@@ -3,11 +3,10 @@ package team.y2k2.globa.api.model.entity;
 import com.google.gson.annotations.SerializedName;
 
 public class Keyword {
-
     @SerializedName("word")
     private String word;
     @SerializedName("importance")
-    private int importance;
+    private float importance;
 
     public String getWord() {
         return word;
@@ -16,9 +15,10 @@ public class Keyword {
         this.word = word;
     }
 
-    public int getImportance() {
+    public float getImportance() {
         return importance;
     }
+
     public void setImportance(int importance) {
         this.importance = importance;
     }
